clean up prompts and imports
