+++ test_rag.py
@@ -1,56 +1,14 @@
 import os
 from rag import RAG
 import adalflow as adal
-from adalflow.core.db import LocalDB
-from adalflow.core.types import Document
 from adalflow.utils import get_adalflow_default_root_path
-from adalflow.components.data_process import TextSplitter, ToEmbeddings
 import tempfile
-
-def create_sample_documents():
-    """Create some sample documents for testing."""
-    sample_texts = [
-        """Alice is a software engineer who loves coding in Python. 
-        She specializes in machine learning and has worked on several NLP projects.
-        Her favorite project was building a chatbot for customer service.""",
-        
-        """Bob is a data scientist with expertise in deep learning.
-        He has published papers on transformer architectures and attention mechanisms.
-        Recently, he's been working on improving RAG systems.""",
-        
-        """The company cafeteria serves amazing tacos on Tuesdays.
-        They also have a great coffee machine that makes perfect lattes.
-        Many employees enjoy their lunch breaks in the outdoor seating area."""
-    ]
-    
-    return [Document(text=text, meta_data={"title": f"doc_{i}"}) 
-            for i, text in enumerate(sample_texts)]
+from data_pipeline import create_sample_documents, transform_documents_and_save_to_db
 
 def initialize_test_database(db_path: str):
     """Initialize database with sample documents."""
     documents = create_sample_documents()
-    
-    data_transformer = adal.Sequential(
-        TextSplitter(split_by="word", chunk_size=50, chunk_overlap=20),
-        ToEmbeddings(
-            embedder=adal.Embedder(
-                model_client=adal.OpenAIClient(),
-                model_kwargs={
-                    "model": "text-embedding-3-small",
-                    "dimensions": 256,
-                    "encoding_format": "float",
-                }
-            ),
-            batch_size=100
-        )
-    )
-    
-    db = LocalDB("test_db")
-    db.register_transformer(transformer=data_transformer, key="split_and_embed")
-    db.load(documents)
-    db.transform(key="split_and_embed")
-    os.makedirs(os.path.dirname(db_path), exist_ok=True)
-    db.save_state(filepath=db_path)
+    transform_documents_and_save_to_db(documents, db_path)
     return db_path
 
 def main():
